+++ src/shared/math/vector3.cpp
@@ -24,7 +24,7 @@ void vectoangles2(const vec3_t& value1, vec3_t& angles)
     }
     else {
         if (value1.xyz[0])
-            yaw = atan2(value1.xyz[1], value1.xyz[0]) * 180 / M_PI;
+            yaw = std::atan2f(value1.xyz[1], value1.xyz[0]) * 180.f / M_PI;
         else if (value1.xyz[1] > 0)
             yaw = 90;
         else
@@ -34,7 +34,7 @@ void vectoangles2(const vec3_t& value1, vec3_t& angles)
             yaw += 360;
 
         forward = sqrt(value1.xyz[0] * value1.xyz[0] + value1.xyz[1] * value1.xyz[1]);
-        pitch = atan2(value1.xyz[2], forward) * 180 / M_PI;
+        pitch = std::atan2f(value1.xyz[2], forward) * 180.f / M_PI;
         if (pitch < 0)
             pitch += 360;
     }
